+++ src/JMS/Serializer/GenericSerializationVisitor.php
@@ -29,7 +29,7 @@ abstract class GenericSerializationVisitor extends AbstractVisitor
     private $dataStack;
     private $data;
 
-    public function setNavigator(GraphNavigator $navigator)
+    public function setNavigator(GraphNavigatorInterface $navigator)
     {
         $this->navigator = $navigator;
         $this->root = null;