+++ src/JMS/Serializer/XmlDeserializationVisitor.php
@@ -42,7 +42,7 @@ class XmlDeserializationVisitor extends AbstractVisitor
         $this->disableExternalEntities = false;
     }
 
-    public function setNavigator(GraphNavigator $navigator)
+    public function setNavigator(GraphNavigatorInterface $navigator)
     {
         $this->navigator = $navigator;
         $this->objectStack = new \SplStack;